backend/requirements.txt
# Core Django
Django==4.2.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
psycopg2-binary==2.9.9

# FastAPI
fastapi==0.110.0
uvicorn[standard]==0.27.0
websockets==12.0
python-multipart==0.0.6

# AI/ML Libraries
langchain==0.1.6
langgraph==0.0.20
ollama==0.1.6

# HTTP Clients
httpx>=0.25.0
requests==2.31.0

# Redis
redis==5.0.1
hiredis==2.3.2

# Environment
python-dotenv==1.0.0

# Audio Processing (for local STT if needed)
# faster-whisper==0.10.0  # Uncomment if using local Whisper

# Utilities
pydantic==2.5.3
pydantic-settings==2.1.0
